Atualização do projeto, acrescentando cadastro de produtos e listar produtos

+++ src/main/java/com/example/auth/domain/product/ProductRequestDTO.java
@@ -3,13 +3,13 @@ package com.example.auth.domain.product;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 public record ProductRequestDTO(
         @NotBlank
         String name,
 
         @NotNull
-        Integer price
+        BigDecimal price
 ) {
 }
