+++ src/main/java/com/example/auth/domain/product/ProductResponseDTO.java
@@ -1,6 +1,8 @@
 package com.example.auth.domain.product;
 
-public record ProductResponseDTO(String id, String name, Integer price) {
+import java.math.BigDecimal;
+
+public record ProductResponseDTO(String id, String name, BigDecimal price) {
     public ProductResponseDTO(Product product){
         this(product.getId(), product.getName(), product.getPrice());
     }
